+++ app/src/main/java/com/quran/labs/androidquran/di/component/activity/PagerActivityComponent.kt
@@ -5,7 +5,9 @@ import com.quran.data.di.ActivityScope
 import com.quran.labs.androidquran.di.component.fragment.QuranPageComponent
 import com.quran.labs.androidquran.di.module.activity.PagerActivityModule
 import com.quran.labs.androidquran.ui.PagerActivity
+import com.quran.labs.androidquran.ui.fragment.AyahPlaybackFragment
 import com.quran.labs.androidquran.ui.fragment.AyahTranslationFragment
+import com.quran.labs.androidquran.ui.fragment.TagBookmarkFragment
 import com.quran.page.common.toolbar.AyahToolBar
 import com.squareup.anvil.annotations.MergeSubcomponent
 import dagger.Subcomponent
@@ -17,8 +19,11 @@ interface PagerActivityComponent {
   fun quranPageComponentBuilder(): QuranPageComponent.Builder
 
   fun inject(pagerActivity: PagerActivity)
-  fun inject(ayahTranslationFragment: AyahTranslationFragment)
   fun inject(ayahToolBar: AyahToolBar)
+
+  fun inject(tagBookmarkFragment: TagBookmarkFragment)
+  fun inject(ayahPlaybackFragment: AyahPlaybackFragment)
+  fun inject(ayahTranslationFragment: AyahTranslationFragment)
 
   @Subcomponent.Builder
   interface Builder {
